requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML - Object Detection
ultralytics>=8.1.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0

# Tracking & Processing
numpy>=1.26.0
scipy>=1.12.0
scikit-image>=0.22.0
Pillow>=10.2.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# LLM API
google-generativeai>=0.3.2

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# API & Validation
pydantic[email]>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# CORS & Security
fastapi-cors>=0.0.6

# Data Export
openpyxl>=3.1.0
pandas>=2.2.0

# Plotting & Analytics
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=8.0.0
httpx>=0.26.0

# Development
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.8.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Optional: GPU Support (uncomment if needed)
# torch-cuda-11-8==1.0.0
# torch-cuda-12-1==1.0.0

# Optional: Database (if using PostgreSQL instead of SQLite)
# psycopg2-binary==2.9.9
# asyncpg==0.29.0

# Optional: Redis (for caching/queue)
# redis==5.0.1
# celery==5.3.4
